Make emoticon endpoint https, emote id is now gint64 and minor formatting

+++ src/gt-twitch.h
@@ -63,7 +63,7 @@ typedef struct
 
 typedef struct
 {
-    gint id;
+    gint64 id;
     gchar* code;
     gint set;
     GdkPixbuf* pixbuf;
